imporved

+++ teardown.sh
@@ -3,10 +3,15 @@
 # Source  environment variables
 . $1/jenkins.env.sh
 
+if [ ! -d "$TEST_FOLDER/$TEST_NAME" ]; then
+  # Control will enter here if $DIRECTORY doesn't exist.
+  exit 0
+fi
+
 # Goto test folder
 cd $TEST_FOLDER/$TEST_NAME
 
-# Remove test container and  node modules volume
+# Test container are stoped and removed if created
 docker-compose stop
 docker-compose rm -f
 # docker volume rm $NAME"test_node_modules"  #Exclude for a faster builds
